requirements.txt
fastmcp
boto3
httpx
google-adk
python-dotenv
